reorganize tests by command type grouping
