+++ modules/tests/test/src/test/scala/com/github/scoquelin/arugula/LettuceRedisKeyAsyncCommandsSpec.scala
@@ -1,0 +1,123 @@
+package com.github.scoquelin.arugula
+
+import scala.concurrent.duration.{DurationInt, DurationLong}
+
+import io.lettuce.core.RedisFuture
+import org.mockito.ArgumentMatchers.{anyLong, anyString}
+import org.mockito.Mockito.{verify, when}
+import org.scalatest.matchers.must.Matchers
+import org.scalatest.{FutureOutcome, wordspec}
+
+class LettuceRedisKeyAsyncCommandsSpec extends wordspec.FixtureAsyncWordSpec with Matchers {
+
+  override type FixtureParam = LettuceRedisCommandsClientFixture.TestContext
+
+  override def withFixture(test: OneArgAsyncTest): FutureOutcome =
+    withFixture(test.toNoArgAsyncTest(new LettuceRedisCommandsClientFixture.TestContext))
+
+  "LettuceRedisKeyAsyncCommands" should {
+    "delegate DEL command to Lettuce and lift result into a Future" in { testContext =>
+      import testContext._
+
+      val expectedValue = 1L
+      val mockRedisFuture: RedisFuture[java.lang.Long] = mockRedisFutureToReturn(expectedValue)
+      when(lettuceAsyncCommands.del(anyString)).thenReturn(mockRedisFuture)
+
+      testClass.del("key").map { result =>
+        result mustBe expectedValue
+        verify(lettuceAsyncCommands).del("key")
+        succeed
+      }
+    }
+
+    "delegate DEL command with multiple keys to Lettuce and lift result into a Future" in { testContext =>
+      import testContext._
+
+      val expectedValue = 2L
+      val mockRedisFuture: RedisFuture[java.lang.Long] = mockRedisFutureToReturn(expectedValue)
+      when(lettuceAsyncCommands.del(anyString, anyString)).thenReturn(mockRedisFuture)
+
+      testClass.del("key1", "key2").map { result =>
+        result mustBe expectedValue
+        verify(lettuceAsyncCommands).del("key1", "key2")
+        succeed
+      }
+    }
+
+
+    "delegate EXISTS command to Lettuce and lift result into a Future" in { testContext =>
+      import testContext._
+
+      val expectedValue = 1L
+      val mockRedisFuture: RedisFuture[java.lang.Long] = mockRedisFutureToReturn(expectedValue)
+      when(lettuceAsyncCommands.exists(anyString)).thenReturn(mockRedisFuture)
+
+      testClass.exists("key").map { result =>
+        result mustBe true
+        verify(lettuceAsyncCommands).exists("key")
+        succeed
+      }
+    }
+
+    "delegate EXISTS command with multiple keys to Lettuce and lift result into a Future" in { testContext =>
+      import testContext._
+
+      val expectedValue = 2L
+      val mockRedisFuture: RedisFuture[java.lang.Long] = mockRedisFutureToReturn(expectedValue)
+      when(lettuceAsyncCommands.exists(anyString, anyString)).thenReturn(mockRedisFuture)
+
+      testClass.exists("key1", "key2").map { result =>
+        result mustBe true
+        verify(lettuceAsyncCommands).exists("key1", "key2")
+        succeed
+      }
+    }
+
+    "delegate expire to PEXPIRE when provided unit is MILLIS, MICROS or NANOS and lift result into a Future" in { testContext =>
+      import testContext._
+
+      val expectedValue = true
+      val mockRedisFuture: RedisFuture[java.lang.Boolean] = mockRedisFutureToReturn(expectedValue)
+      when(lettuceAsyncCommands.pexpire(anyString, anyLong)).thenReturn(mockRedisFuture)
+
+      testClass.expire("key", 1000.microseconds).map { result =>
+        result mustBe expectedValue
+        verify(lettuceAsyncCommands).pexpire("key", 1L)
+        succeed
+      }
+    }
+
+    "delegate expire to EXPIRE when provided unit is SECONDS or higher and lift result into a Future" in { testContext =>
+      import testContext._
+
+      val expectedValue = true
+      val mockRedisFuture: RedisFuture[java.lang.Boolean] = mockRedisFutureToReturn(expectedValue)
+      when(lettuceAsyncCommands.expire(anyString, anyLong)).thenReturn(mockRedisFuture)
+
+      testClass.expire("key", 1.minute).map { result =>
+        result mustBe expectedValue
+        verify(lettuceAsyncCommands).expire("key", 60L)
+        succeed
+      }
+    }
+
+    "delegate TTL command to Lettuce and lift result into a Future" in { testContext =>
+      import testContext._
+
+      val expectedValue = 1L
+      val mockRedisFuture: RedisFuture[java.lang.Long] = mockRedisFutureToReturn(expectedValue)
+      when(lettuceAsyncCommands.exists(anyString, anyString)).thenReturn(mockRedisFuture)
+
+      when(lettuceAsyncCommands.ttl(anyString)).thenReturn(mockRedisFuture)
+      testClass.ttl("key").map {
+        case Some(result) =>
+          result mustBe (expectedValue).seconds
+          verify(lettuceAsyncCommands).ttl("key")
+          succeed
+        case None =>
+          fail(s"TTL for key should be $expectedValue second(s)")
+      }
+    }
+  }
+
+}
